[Decomp][Player] Cohesion split: extract duel lifecycle to PlayerDuel.cpp
