+++ src/game/Object/Player.cpp
@@ -4727,177 +4727,6 @@ void Player::UpdateZone(uint32 newZone, uint32 newArea)
     UpdateZoneDependentPets();
 }
 
-// If players are too far way of duel flag... then player loose the duel
-void Player::CheckDuelDistance(time_t currTime)
-{
-    if (!duel)
-    {
-        return;
-    }
-
-    GameObject* obj = GetMap()->GetGameObject(GetGuidValue(PLAYER_DUEL_ARBITER));
-    if (!obj)
-    {
-        // player not at duel start map
-        DuelComplete(DUEL_FLED);
-        return;
-    }
-
-    if (duel->outOfBound == 0)
-    {
-        if (!IsWithinDistInMap(obj, 50))
-        {
-            duel->outOfBound = currTime;
-
-            WorldPacket data(SMSG_DUEL_OUTOFBOUNDS, 0);
-            GetSession()->SendPacket(&data);
-        }
-    }
-    else
-    {
-        if (IsWithinDistInMap(obj, 40))
-        {
-            duel->outOfBound = 0;
-
-            WorldPacket data(SMSG_DUEL_INBOUNDS, 0);
-            GetSession()->SendPacket(&data);
-        }
-        else if (currTime >= (duel->outOfBound + 10))
-        {
-            DuelComplete(DUEL_FLED);
-        }
-    }
-}
-
-/**
- * @brief Finalizes an active duel and cleans up all duel state.
- *
- * @param type The duel completion result.
- */
-void Player::DuelComplete(DuelCompleteType type)
-{
-    // duel not requested
-    if (!duel)
-    {
-        return;
-    }
-
-    Player* opponent = duel->opponent;
-
-    // opponent may have been removed (disconnect during duel)
-    if (!opponent)
-    {
-        // Remove Duel Flag object
-        if (GameObject* obj = GetMap()->GetGameObject(GetGuidValue(PLAYER_DUEL_ARBITER)))
-        {
-            duel->initiator->RemoveGameObject(obj, true);
-        }
-
-        SetGuidValue(PLAYER_DUEL_ARBITER, ObjectGuid());
-        SetUInt32Value(PLAYER_DUEL_TEAM, 0);
-        delete duel;
-        duel = NULL;
-        return;
-    }
-
-    WorldPacket data(SMSG_DUEL_COMPLETE, (1));
-    data << (uint8)((type != DUEL_INTERRUPTED) ? 1 : 0);
-    GetSession()->SendPacket(&data);
-    opponent->GetSession()->SendPacket(&data);
-
-    if (type != DUEL_INTERRUPTED)
-    {
-        data.Initialize(SMSG_DUEL_WINNER, (1 + 20));          // we guess size
-        data << (uint8)((type == DUEL_WON) ? 0 : 1);          // 0 = just won; 1 = fled
-        data << opponent->GetName();
-        data << GetName();
-        SendMessageToSet(&data, true);
-    }
-
-    // Used by Eluna
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = GetEluna())
-    {
-        e->OnDuelEnd(opponent, this, type);
-    }
-#endif /* ENABLE_ELUNA */
-
-    if (type == DUEL_WON)
-    {
-        GetAchievementMgr().UpdateAchievementCriteria(ACHIEVEMENT_CRITERIA_TYPE_LOSE_DUEL, 1);
-        opponent->GetAchievementMgr().UpdateAchievementCriteria(ACHIEVEMENT_CRITERIA_TYPE_WIN_DUEL, 1);
-    }
-
-    // Remove Duel Flag object
-    if (GameObject* obj = GetMap()->GetGameObject(GetGuidValue(PLAYER_DUEL_ARBITER)))
-    {
-        duel->initiator->RemoveGameObject(obj, true);
-    }
-
-    /* remove auras */
-    // TODO: Needs a simpler method
-    std::vector<uint32> auras2remove;
-    SpellAuraHolderMap const& vAuras = opponent->GetSpellAuraHolderMap();
-    for (SpellAuraHolderMap::const_iterator i = vAuras.begin(); i != vAuras.end(); ++i)
-    {
-        SpellAuraHolder const* aura = i->second;
-        if (!aura->IsPositive() && aura->GetCasterGuid() == GetObjectGuid() && aura->GetAuraApplyTime() >= duel->startTime)
-        {
-            auras2remove.push_back(aura->GetId());
-        }
-    }
-
-    for (size_t i = 0; i < auras2remove.size(); ++i)
-    {
-        opponent->RemoveAurasDueToSpell(auras2remove[i]);
-    }
-
-    auras2remove.clear();
-    SpellAuraHolderMap const& auras = GetSpellAuraHolderMap();
-    for (SpellAuraHolderMap::const_iterator i = auras.begin(); i != auras.end(); ++i)
-    {
-        SpellAuraHolder const* aura = i->second;
-        if (!aura->IsPositive() && aura->GetCasterGuid() == opponent->GetObjectGuid() && aura->GetAuraApplyTime() >= duel->startTime)
-        {
-            auras2remove.push_back(aura->GetId());
-        }
-    }
-    for (size_t i = 0; i < auras2remove.size(); ++i)
-    {
-        RemoveAurasDueToSpell(auras2remove[i]);
-    }
-
-    // cleanup combo points
-    if (GetComboTargetGuid() == opponent->GetObjectGuid())
-    {
-        ClearComboPoints();
-    }
-    else if (GetComboTargetGuid() == opponent->GetPetGuid())
-    {
-        ClearComboPoints();
-    }
-
-    if (opponent->GetComboTargetGuid() == GetObjectGuid())
-    {
-        opponent->ClearComboPoints();
-    }
-    else if (opponent->GetComboTargetGuid() == GetPetGuid())
-    {
-        opponent->ClearComboPoints();
-    }
-
-    // cleanups
-    SetGuidValue(PLAYER_DUEL_ARBITER, ObjectGuid());
-    SetUInt32Value(PLAYER_DUEL_TEAM, 0);
-    opponent->SetGuidValue(PLAYER_DUEL_ARBITER, ObjectGuid());
-    opponent->SetUInt32Value(PLAYER_DUEL_TEAM, 0);
-
-    delete opponent->duel;
-    opponent->duel = NULL;
-    delete duel;
-    duel = NULL;
-}
-
 /**
  * @brief Sends a single world state update to the client.
  *
@@ -8868,13 +8697,6 @@ void Player::SendTimeSync()
     // Schedule next sync in 10 sec
     m_timeSyncTimer = 10000;
     m_timeSyncServer = GameTime::GetGameTimeMS();
-}
-
-void Player::SendDuelCountdown(uint32 counter)
-{
-    WorldPacket data(SMSG_DUEL_COUNTDOWN, 4);
-    data << uint32(counter);                                // seconds
-    GetSession()->SendPacket(&data);
 }
 
 bool Player::IsImmuneToSpellEffect(SpellEntry const* spellInfo, SpellEffectIndex index, bool castOnSelf) const
